requirements.txt
flask
opencv-python-headless
mediapipe
numpy
Pillow